+++ map/framework.hpp
@@ -278,7 +278,7 @@ public:
   void PrepareSearch(bool hasPt, double lat = 0.0, double lon = 0.0);
   bool Search(search::SearchParams const & params);
   bool GetCurrentPosition(double & lat, double & lon) const;
-  void ShowSearchResult(search::Result const & res, size_t index);
+  void ShowSearchResult(search::Result const & res, int index);
 
   /// Calculate distance and direction to POI for the given position.
   /// @param[in]  point             POI's position;
